Atualizações

+++ src/os/index.php
@@ -8,7 +8,9 @@
         $fotos = [];
         while($d = mysqli_fetch_object($result)){
             $fotos[] = $d->foto;
-            echo "<p><img src='src/os/fotos/{$_POST['excluir']}/{$d->foto}' width='100' /></p>";
+            if(is_file("fotos/{$_POST['excluir']}/{$d->foto}")){
+                echo "<p><img src='src/os/fotos/{$_POST['excluir']}/{$d->foto}' width='100' /></p>";
+            }
         }
         
         $query = "delete from os where codigo = '{$_POST['excluir']}'";
